VRC-4 - navigation between login and register
