+++ app/src/main/java/com/example/vrach/MainActivity.kt
@@ -10,7 +10,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.vrach.ui.VrachApp
 import com.example.vrach.ui.theme.VrachTheme
 
 class MainActivity : ComponentActivity() {
