+++ app/src/main/java/com/example/vrach/ui/login/LoginOptionScreen.kt
@@ -1,44 +1,30 @@
 package com.example.vrach.ui.login
 
-import android.graphics.drawable.DrawableContainer
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.magnifier
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.ClickableText
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
 import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonColors
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.modifier.modifierLocalConsumer
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.TextStyle
@@ -48,13 +34,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.vrach.R
-import com.example.vrach.ui.theme.PurpleGrey80
-import com.example.vrach.ui.theme.PurpleWhite
 import com.example.vrach.ui.util.clickableWithoutRipple
-import java.text.DecimalFormat
 
 @Composable
-fun LoginOptionScreen() {
+fun LoginOptionScreen(
+    onSignUpClicked: () -> Unit = {}
+) {
     val scrollState = rememberScrollState()
     val interactionSource = remember { MutableInteractionSource() }
     Column(
@@ -101,9 +86,7 @@ fun LoginOptionScreen() {
                 horizontalArrangement = Arrangement.Center,
                 modifier = Modifier.clickableWithoutRipple(
                     interactionSource = interactionSource,
-                    onClick = {
-
-                    }
+                    onClick = onSignUpClicked
                 )
             ) {
                 Text(
@@ -180,5 +163,5 @@ fun LoginButtons(
 @Preview(showBackground = true)
 @Composable
 fun VrachPreview() {
-    LoginOptionScreen()
+ //   LoginOptionScreen(navController = (Int) -> Unit)
 }
